fix jc test

+++ pa/src/test/scala/com/adatao/pa/spark/execution/GraphSuite.scala
@@ -156,7 +156,7 @@ class GraphSuite extends ABigRClientTest {
 
     val cosineResult2 = cosineResult.map {
       row => row.replace("\"", "").split("\\s+")
-    }.map{arr => if(arr.size == 3) Array(arr(0), arr(1), arr(2)) else Array()}
+    }.map{arr => if(arr.size == 3) Array(arr(0), arr(1), arr(2).toDouble) else Array()}
 
     assert(cosineResult2.size == 1)
     assert(cosineResult2(0)(0).asInstanceOf[String] == "-3351804022671213759")
